pyproject.toml
[project]
name = "personal-health-coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "google-auth>=2.40.1",
    "ollama>=0.4.8",
    "psycopg[binary]>=3.2.9",
    "pydantic-settings>=2.9.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
]

[tool.mypy]
explicit_package_bases = true

[tool.ruff]
exclude = [
    "backend/alembic",  
]
line-length = 120